students/K33392/practical_works/stukalov_artem/simple_drf_project/pyproject.toml
[tool.poetry]
name = "simple-drf-project"
version = "0.1.0"
description = ""
authors = ["TUTOR03 <stukalov.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
